+++ src/h5web/vis-packs/core/containers/index.ts
@@ -3,4 +3,4 @@ export { default as ScalarVisContainer } from './ScalarVisContainer';
 export { default as MatrixVisContainer } from './MatrixVisContainer';
 export { default as LineVisContainer } from './LineVisContainer';
 export { default as HeatmapVisContainer } from './HeatmapVisContainer';
-export { default as PhaseMapVisContainer } from './PhaseMapVisContainer';
+export { default as ComplexVisContainer } from './ComplexVisContainer';
